made lots of fixes

+++ src/main.cpp
@@ -19,9 +19,9 @@
 
 #include "NN91_Basic.h"
 
-const size_t LOSS_VAL = -1;
-const size_t DRAW_VAL = 0;
-const size_t WIN_VAL = 2;
+const int8_t LOSS_VAL = -1;
+const int8_t DRAW_VAL = 0;
+const int8_t WIN_VAL = 2;
 const std::vector<int> NET_SIZE {5,4,3};
 
 struct NetTracker {
